backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
groq==0.9.0
python-dotenv==1.0.1
